changed lastIndex from int to long...sorry new client db
git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS12]$/geosyncronization/Geosynk/Development;C40614

+++ Kartverket.Geosynkronisering.Subscriber/Kartverket.Geosynkronisering.Subscriber.DL/SubscriberDataset.cs
@@ -9,7 +9,7 @@ namespace Kartverket.Geosynkronisering.Subscriber.DL
     {
         public int DatasetId { get; set; }
         public string Name { get; set; }
-        public int LastIndex { get; set; }
+        public long LastIndex { get; set; }
         public string SynchronizationUrl { get; set; }
         public string ClientWfsUrl { get; set; }
         public int MaxCount { get; set; }
